Add $mode variable to control the cron file permissions
Note that I've changed the default file permissions to 0644, which I
consider to be more widley appropriate than the 0640 which was
previously being used, but this modification allows that to be overriden
by default.

It'd be nice to be able to set default permissions, but that is not
implemented here.