store file mode as a string
This allows for puppet4 support.

"The mode attribute of a file resource must be a string. If you use an
actual octal number, it will be converted to a decimal number, then
converted back to a string representing the wrong number when it comes
time to run the chown command."

https://docs.puppetlabs.com/puppet/latest/reference/experiments_future.html#check-the-mode-attribute-of-any-file-resources